+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Faro shuffle".
 - Tests for "Morellet’s random lines".
 - Tests for "Decode the message".
+- Tests for "Detective geek".
 
 ## [1.9.0] - 2022-06-01
 ### Added